# apps/api-svc/internal/repository/postgres/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package postgres

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Camp struct {
	ID        pgtype.UUID        `json:"id"`
	Name      string             `json:"name"`
	Sport     string             `json:"sport"`
	Location  string             `json:"location"`
	Capacity  int32              `json:"capacity"`
	StartDate pgtype.Date        `json:"start_date"`
	EndDate   pgtype.Date        `json:"end_date"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}